Added flag to simply lock the machine
This is only for OSX

+++ check.sh
@@ -8,10 +8,17 @@ fi
 
 while [ -d "$1" ]; do
 	# Don't do anything
-	sleep 4
+	sleep 2
 done
 
 echo "Drive removed. Shutting down.";
+
+if [ $2 = "-l" ]
+then
+	/System/Library/CoreServices/Menu\ Extras/User.menu/Contents/Resources/CGSession -suspend
+	exit 0
+
+fi
 
 shutdown -r now
